+++ src/components/app/App.tsx
@@ -1,0 +1,14 @@
+import React from 'react';
+import './app.css';
+
+import { CountryGraphs } from '../country-graphs/country-graphs';
+
+function App() {
+  return (
+    <div className="app">
+      <CountryGraphs />
+    </div>
+  );
+}
+
+export default App;
