+++ src/components/app/app.css
@@ -1,0 +1,3 @@
+.app {
+  padding: 16px;
+}
